@@ -3,12 +3,12 @@ import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 buildscript {
   ext {
-    kotlin_version = '1.8.10'
+    kotlin_version = '2.1.21'
     workflow_version = "1.12.1-beta04"
 
     deps = [
         activityktx: 'androidx.activity:activity-ktx:1.3.0',
-        agp: "com.android.tools.build:gradle:8.3.2",
+        agp: "com.android.tools.build:gradle:8.8.0",
         appcompat: 'androidx.appcompat:appcompat:1.3.1',
         constraintlayout: 'androidx.constraintlayout:constraintlayout:2.0.1',
         kotlin: [
