@@ -46,7 +46,7 @@ import kotlin.reflect.typeOf
  *
  * See [renderChild].
  */
-public interface BaseRenderContext<out PropsT, StateT, in OutputT> {
+public interface BaseRenderContext<PropsT, StateT, OutputT> {
 
   public val runtimeConfig: RuntimeConfig
 
